share config json file with the update app
some cpld configurations require access to multiple i2c busses.
it is futile to try to individually communicate all the necessary
parameters, so share the path to the config file to the update app
so it can figure out what it needs.

Signed-off-by: Radivoje (Ogi) Jovanovic <rjovanovic@nvidia.com>
fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-3884