fungibility: Modify inventory path for target filter.
GPU inventory object was changed to
/xyz/openbmc_project/inventory/system/chassis/HGX_PCIeRetimer_
and item updater was relying on this object path to validate the target.

This change was breaking the target filering logic in item updater for
retimer.Also Model property lookup was not required since only validation
of path is required.

Unit Tests:
  1. Verify retimer target filter update
  2. Verify firmware update without target filters
  3. Verify retimer update with targets as empty, and pldm, non-pldm devices

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-5032

Signed-off-by: Chinmay Shripad Hegde <chinmays@nvidia.com>