fix fusing status reporting for switchtec switch
the switch is only available on i2c bus when the platform is ON.
add match that will fire when the platform goes ON to check the
fusing status.
also do not start the service until the mux is present on the system.
the mux is present after entity-manager has instantiated it

Signed-off-by: Radivoje (Ogi) Jovanovic <rjovanovic@nvidia.com>
fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-7662