fw-status: Removing dependency on dbusutils for fetching service
The utility to fetch service name in dbusutils has multiple retries and
sleeps for 1 second between each retry. This is causing the service to
hang up and unable to process any MCTP responses.

Removed the dependency on dbusutils and created a similar function to
fetch the service name

Fixes NVBug https://

Signed-off-by: P Arun Kumar Reddy parunkumarr@nvidia.com