Add "Enabled" property check to MCTP endpoint enumeration
Makes debug token updater check whether an MCTP endpoint is enabled or
disabled. When an endpoint is disabled, it does not respond to MCTP VDM
requests, resulting in a timeout. With multiple endpoints in a disabled
state, a timeout of 180 seconds was being triggered in PLDM.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4943229

Signed-off-by: Piotr Juda <pjuda@nvidia.com>