fw-status: Changing the Object Manager to '/'
The previous ObjectManager for the service had a trailing '/' which
caused it to not populate the ObjectManager correctly.

On boot up, when the ObjectMapper and FWStatus service start up, the
FWStatus does not publish any D-Bus objects (including the
ObjectManager) after requesting a service name from D-Bus due to its
dependency on EntityManager, and MCTP services. This causes the
ObjectMapper to ignore the service, leading to empty responses for
ObjectMapper calls, when fetching the chassis D-Bus objects for
BootStatus

Tests:
- BootStatusCode is populated at RoT Protected components

Fixes NVBug https://nvbugspro.nvidia.com/bug/5014715

Signed-off-by: P Arun Kumar Reddy parunkumarr@nvidia.com