Prevent creation of existing DBus object
The FWStatus service lacked safeguards to prevent creating a
DBus object with an identical object path. This caused crashes
when the ConfigurableStateManager was restarted, as the service
attempted to register an already existing object.

This update ensures that attempts to create an existing DBus
object are intercepted, preventing service crashes.

Unit Tests:
1. Restart ConfigurableStateManager multiple times using the
   following command:

   systemctl restart xyz.openbmc_project.State.ConfigurableStateManager

2. Confirm the FWStatus service does not crash. Instead,
   journalctl logs should display appropriate warnings when
   attempting to create an existing DBus object.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4978333

Signed-off-by: Robert Papros <rpapros@nvidia.com>