change the code signature check is done
signature check is being enabled. However, the code needs to be backward
compatible with the current implementation  that does not have the
signature checking. Due to this customers requsted to have an option
to turn on/off the signature checking for a few releases.
Due to this request the code is changed to accomodate another
change in which the siganture check will be dynamic

Signed-off-by: Radivoje (Ogi) Jovanovic <rjovanovic@nvidia.com>
fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-16387