REVERTME: tie the signature check to the presence of the file
compatible with the current implementation  that does not have the
signature checking. Due to this customers requsted to have an option
to turn on/off the signature checking for a few releases.
with this change the code will only do signature check if there is a
file at the certain location.
the reason why it is done with the file instead of dbus is because
this is temporary and will be removed as soon as possible.
doing it over dbus would be significantly bigger change and
exposing it over redfish would cause mockup change etc.
keep it simple and remove.

Signed-off-by: Radivoje (Ogi) Jovanovic <rjovanovic@nvidia.com>
fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-16387