MCU Recovery: Add additional check for Iris MCU
Reason:
For general case, we can determine the MCU is healthy if the functional
PID matches the it's current product ID. However, for Iris MCU (PX86E)
it's different as Iris MCU PIDs are the same when it's in ISP mode.
So we need to check if the device has MCTP class.

Tested:
* Confirmed that mcu-recovery-tool can correctly report MCU's heath
  status
* Performed the SMA FW recovery on GB300 Bianca

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-16602

Signed-off-by: Harry Sung <harrys@nvidia.com>